Clear the mechanical strict-clippy findings (B2)
Behavior-preserving rewrites for strict clippy (223 -> 190 errors on
67d91f54): collapsible ifs become let-chains; a `?` replaces a
return-None else; a same-type u64 cast, a needless return, and a lazy
None closure go; match scrutinee blocks move into a let; a byte string
replaces a byte array; a doc list gets its separating blank line; an
irrefutable single-variant if-let becomes a let; the Unix accept loop is
an async fn; the zero-sized OwnerPermit is consumed by destructuring
instead of drop(); ClientCleanupFault becomes PoisonedCleanupLock
{Plane, Connection, Router, Mailbox, Pool} (names are not observable);
StoredSessionContext, JsonProblem, and KeyFail become pub(crate) to
match the items that expose them; two unused imports go.

Left for the event-driven writer's files: data_plane/driver.rs and
daemon/control/managed_git.rs (one each), and an owner_loop test import.

+++ src/recovery/state_directory.rs
@@ -423,9 +423,9 @@ impl StateDirectoryOwnership {
         FileType::from_raw_mode(current.st_mode) == FileType::Directory
             && FileType::from_raw_mode(retained.st_mode) == FileType::Directory
             && current.st_dev as u64 == self.0.device
-            && current.st_ino as u64 == self.0.inode
+            && current.st_ino == self.0.inode
             && retained.st_dev as u64 == self.0.device
-            && retained.st_ino as u64 == self.0.inode
+            && retained.st_ino == self.0.inode
     }
 
     /// Reject replacement detected before a filesystem operation.
